remove model.half() for ipex summarizer

+++ smart-classroom/components/llm/ipex/summarizer.py
@@ -59,7 +59,7 @@ class Summarizer(BaseSummarizer):
             model_hub=model_hub
         )
         self.device = device
-        self.model = self.model.half().to(self.device)
+        self.model = self.model.to(self.device)
 
         self.temperature = temperature
